Validated AircareApi handling. Progress in Event notifications on Security side

+++ bundles/org.openhab.binding.netatmo/src/main/java/org/openhab/binding/netatmo/internal/api/dto/NADashboard.java
@@ -179,7 +179,7 @@ public class NADashboard {
         return gustStrength;
     }
 
-    public double getHealthIdx() {
+    public int getHealthIdx() {
         return healthIdx;
     }
 }
